+++ resources/release-notes.md
@@ -2,8 +2,9 @@
 
 * Properly include `libintl` such that for example `:languge` works.
 * Dependencies updates:
-  - sparkle-project/Sparkle@2.5.2
-
+  - sparkle-project/Sparkle@2.6.0
+  - Quick/Nimble@13.3.0
+  - apple/swift-collections@1.1.0
 
 # v0.46.0-20240102.233758